requirements.txt
docker>=7.0.0
Flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
PyYAML>=6.0.1